requirements.txt
discord.py==2.7.3
python-dotenv==1.0.1
